Refactoring. Main menu handlers are hooks now.
